requirements.txt
Flask==3.0.0
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
mysql-connector-python==8.2.0
gunicorn==21.2.0
python-dotenv==1.0.0
face-recognition==1.3.0
numpy==1.26.2
